build: reorganise packages of ndk instrumentation
