+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/crash/ndk/NativeCoreModule.kt
@@ -1,0 +1,13 @@
+package io.embrace.android.embracesdk.internal.instrumentation.crash.ndk
+
+import io.embrace.android.embracesdk.internal.SharedObjectLoader
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.jni.JniDelegate
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.symbols.SymbolService
+
+interface NativeCoreModule {
+    val sharedObjectLoader: SharedObjectLoader
+    val processor: NativeCrashProcessor
+    val symbolService: SymbolService
+    val delegate: JniDelegate
+    val nativeCrashHandlerInstaller: NativeCrashHandlerInstaller?
+}
