+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/crash/ndk/NativeCoreModuleImpl.kt
@@ -1,19 +1,24 @@
-package io.embrace.android.embracesdk.internal.injection
+package io.embrace.android.embracesdk.internal.instrumentation.crash.ndk
 
 import android.os.Build
 import io.embrace.android.embracesdk.internal.SharedObjectLoader
 import io.embrace.android.embracesdk.internal.SharedObjectLoaderImpl
 import io.embrace.android.embracesdk.internal.delivery.storage.StorageLocation
 import io.embrace.android.embracesdk.internal.handler.AndroidMainThreadHandler
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashHandlerInstaller
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashHandlerInstallerImpl
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashProcessor
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashProcessorImpl
-import io.embrace.android.embracesdk.internal.ndk.NativeInstallMessage
-import io.embrace.android.embracesdk.internal.ndk.jni.JniDelegate
-import io.embrace.android.embracesdk.internal.ndk.jni.JniDelegateImpl
-import io.embrace.android.embracesdk.internal.ndk.symbols.SymbolService
-import io.embrace.android.embracesdk.internal.ndk.symbols.SymbolServiceImpl
+import io.embrace.android.embracesdk.internal.injection.ConfigModule
+import io.embrace.android.embracesdk.internal.injection.CoreModule
+import io.embrace.android.embracesdk.internal.injection.EssentialServiceModule
+import io.embrace.android.embracesdk.internal.injection.InitModule
+import io.embrace.android.embracesdk.internal.injection.OpenTelemetryModule
+import io.embrace.android.embracesdk.internal.injection.PayloadSourceModule
+import io.embrace.android.embracesdk.internal.injection.StorageModule
+import io.embrace.android.embracesdk.internal.injection.WorkerThreadModule
+import io.embrace.android.embracesdk.internal.injection.asFile
+import io.embrace.android.embracesdk.internal.injection.singleton
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.jni.JniDelegate
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.jni.JniDelegateImpl
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.symbols.SymbolService
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.symbols.SymbolServiceImpl
 import io.embrace.android.embracesdk.internal.utils.Provider
 import io.embrace.android.embracesdk.internal.utils.Uuid
 import io.embrace.android.embracesdk.internal.worker.Worker
