+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/crash/ndk/NativeCrashProcessor.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.internal.ndk
+package io.embrace.android.embracesdk.internal.instrumentation.crash.ndk
 
 import io.embrace.android.embracesdk.internal.payload.NativeCrashData
