+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/crash/ndk/symbols/SymbolServiceImpl.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.internal.ndk.symbols
+package io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.symbols
 
 import android.util.Base64
 import io.embrace.android.embracesdk.internal.DeviceArchitecture
