+++ embrace-android-core/src/test/java/io/embrace/android/embracesdk/internal/injection/NativeCoreModuleImplTest.kt
@@ -10,6 +10,7 @@ import io.embrace.android.embracesdk.fakes.injection.FakeInitModule
 import io.embrace.android.embracesdk.fakes.injection.FakePayloadSourceModule
 import io.embrace.android.embracesdk.fakes.injection.FakeStorageModule
 import io.embrace.android.embracesdk.fakes.injection.FakeWorkerThreadModule
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.createNativeCoreModule
 import io.mockk.mockk
 import org.junit.Assert.assertNotNull
 import org.junit.Test
