+++ embrace-android-features/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/crash/ndk/NativeCrashDataSource.kt
@@ -1,6 +1,5 @@
 package io.embrace.android.embracesdk.internal.instrumentation.crash.ndk
 
 import io.embrace.android.embracesdk.internal.arch.datasource.DataSource
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashService
 
 internal interface NativeCrashDataSource : DataSource, NativeCrashService
