+++ embrace-android-features/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/crash/ndk/NativeFeatureModule.kt
@@ -1,7 +1,5 @@
 package io.embrace.android.embracesdk.internal.instrumentation.crash.ndk
 
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashService
-
 interface NativeFeatureModule {
     val nativeCrashService: NativeCrashService?
 }
