+++ embrace-android-features/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/crash/ndk/NativeFeatureModuleSupplier.kt
@@ -4,7 +4,6 @@ import io.embrace.android.embracesdk.internal.injection.AndroidServicesModule
 import io.embrace.android.embracesdk.internal.injection.ConfigModule
 import io.embrace.android.embracesdk.internal.injection.InitModule
 import io.embrace.android.embracesdk.internal.injection.InstrumentationModule
-import io.embrace.android.embracesdk.internal.injection.NativeCoreModule
 
 /**
  * Function that returns an instance of [NativeFeatureModule]. Matches the signature of the constructor for [NativeFeatureModuleImpl]
