+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeNativeCrashProcessor.kt
@@ -1,6 +1,6 @@
 package io.embrace.android.embracesdk.fakes
 
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashProcessor
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.NativeCrashProcessor
 import io.embrace.android.embracesdk.internal.payload.NativeCrashData
 
 class FakeNativeCrashProcessor : NativeCrashProcessor {
