+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeNativeFeatureModule.kt
@@ -1,7 +1,7 @@
 package io.embrace.android.embracesdk.fakes
 
 import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.NativeFeatureModule
-import io.embrace.android.embracesdk.internal.ndk.NativeCrashService
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.NativeCrashService
 
 class FakeNativeFeatureModule(
     override val nativeCrashService: NativeCrashService = FakeNativeCrashService(),
