+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeSymbolService.kt
@@ -1,6 +1,6 @@
 package io.embrace.android.embracesdk.fakes
 
-import io.embrace.android.embracesdk.internal.ndk.symbols.SymbolService
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.symbols.SymbolService
 
 class FakeSymbolService(
     override val symbolsForCurrentArch: MutableMap<String, String> = mutableMapOf(),
